+++ tests/Codeception/Acceptance/Basket/PlaceOrderBaseCest.php
@@ -90,17 +90,17 @@ abstract class PlaceOrderBaseCest extends BaseCest
         return $result['data']['basketCreate']['id'];
     }
 
-    protected function addProductToBasket(AcceptanceTester $I, string $basketId, string $productId, float $amount): array
+    protected function addItemToBasket(AcceptanceTester $I, string $basketId, string $productId, float $amount): array
     {
         $variables = [
-            'basketId'  => $basketId,
-            'productId' => $productId,
+            'basketId'  => new ID($basketId),
+            'productId' => new ID($productId),
             'amount'    => $amount,
         ];
 
         $mutation = '
-            mutation ($basketId: String!, $productId: String!, $amount: Float! ) {
-                basketAddProduct(
+            mutation ($basketId: ID!, $productId: ID!, $amount: Float! ) {
+                basketAddItem(
                     basketId: $basketId,
                     productId: $productId,
                     amount: $amount
@@ -118,7 +118,7 @@ abstract class PlaceOrderBaseCest extends BaseCest
 
         $result = $this->getGQLResponse($I, $mutation, $variables);
 
-        return $result['data']['basketAddProduct']['items'];
+        return $result['data']['basketAddItem']['items'];
     }
 
     protected function queryBasketDeliveryMethods(AcceptanceTester $I, string $basketId): array