+++ tests/Codeception/Acceptance/Basket/PlaceOrderCest.php
@@ -31,7 +31,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket
         $basketId = $this->createBasket($I, 'my_anonymous_cart');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 2);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 2);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -53,7 +53,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket
         $basketId = $this->createBasket($I, 'my_cart_one');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 2);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 2);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -90,7 +90,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket
         $basketId = $this->createBasket($I, self::DEFAULT_SAVEDBASKET);
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 2);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 2);
         $this->setBasketDeliveryMethod($I, $basketId, self::TEST_SHIPPING);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_TEST);
 
@@ -126,11 +126,11 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare default basket
         $defaultBasketId = $this->createBasket($I, self::DEFAULT_SAVEDBASKET);
-        $this->addProductToBasket($I, $defaultBasketId, self::PRODUCT_ID, 2);
+        $this->addItemToBasket($I, $defaultBasketId, self::PRODUCT_ID, 2);
 
         //prepare ordering basket
         $basketId = $this->createBasket($I, 'specialBasket');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 2);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 2);
         $this->setBasketDeliveryMethod($I, $basketId, self::TEST_SHIPPING);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_TEST);
 
@@ -148,7 +148,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket
         $basketId = $this->createBasket($I, 'my_cart_two');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 2);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 2);
         $this->setBasketDeliveryAddress($I, $basketId);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
@@ -174,7 +174,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket
         $basketId = $this->createBasket($I, 'my_cart_three');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 2);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 2);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -197,7 +197,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket
         $basketId = $this->createBasket($I, 'my_cart_four');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 2);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 2);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -246,7 +246,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket with invoice address
         $basketId = $this->createBasket($I, 'my_cart_five');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 3);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 3);
         $this->setBasketDeliveryAddress($I, $basketId, self::ALTERNATE_COUNTRY);
 
         //shipping method not supported
@@ -266,7 +266,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket with german delivery address
         $basketId = $this->createBasket($I, 'my_cart_six');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 3);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 3);
         $this->setBasketDeliveryAddress($I, $basketId); //Germany
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
@@ -293,7 +293,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket with invoice address
         $basketId = $this->createBasket($I, 'my_cart_seven');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 3);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 3);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -322,7 +322,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket
         $basketId = $this->createBasket($I, 'cart_with_discount');
-        $this->addProductToBasket($I, $basketId, self::DISCOUNT_PRODUCT, 1);
+        $this->addItemToBasket($I, $basketId, self::DISCOUNT_PRODUCT, 1);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -362,7 +362,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket with invoice address
         $basketId = $this->createBasket($I, 'my_cart_del_cost_flag');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 2);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 2);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -400,7 +400,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket
         $basketId = $this->createBasket($I, 'cart_with_agb_given');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 1);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 1);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -430,7 +430,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket with invoice address
         $basketId = $this->createBasket($I, 'cart_without_agb_given');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 1);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 1);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -457,7 +457,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket with invoice address
         $basketId = $this->createBasket($I, 'cart_with_agb_refused');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 1);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 1);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -484,7 +484,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket with invoice address
         $basketId = $this->createBasket($I, 'cart_with_null_agb');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 1);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 1);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -510,7 +510,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket with invoice address
         $basketId = $this->createBasket($I, 'cart_with_agb_not_required_but_given');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 1);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 1);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -539,7 +539,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket with invoice address
         $basketId = $this->createBasket($I, 'cart_with_agb_not_required_but_refused');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 1);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 1);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -564,7 +564,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket with invoice address
         $basketId = $this->createBasket($I, 'cart_with_agb_not_required_and_not_given');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 1);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 1);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -589,7 +589,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket with invoice address
         $basketId = $this->createBasket($I, 'cart_with_null_agb');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 1);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 1);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -610,7 +610,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket
         $basketId = $this->createBasket($I, 'cart_with_files');
-        $this->addProductToBasket($I, $basketId, self::DOWNLOADABLE_FILE, 1);
+        $this->addItemToBasket($I, $basketId, self::DOWNLOADABLE_FILE, 1);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -641,7 +641,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket
         $basketId = $this->createBasket($I, 'cart_below_min_price');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 1);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 1);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -668,7 +668,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket
         $basketId = $this->createBasket($I, 'cart_with_not_buyable_product');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 5);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 5);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);
 
@@ -696,7 +696,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket
         $basketId = $this->createBasket($I, 'no_delivery_method');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 1);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 1);
 
         //place the order
         $result = $this->placeOrder($I, $basketId);
@@ -711,7 +711,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket
         $basketId = $this->createBasket($I, 'no_payment_method');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 1);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 1);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
 
         //place the order
@@ -767,7 +767,7 @@ final class PlaceOrderCest extends PlaceOrderBaseCest
 
         //prepare basket
         $basketId = $this->createBasket($I, 'not_yet_ordered');
-        $this->addProductToBasket($I, $basketId, self::PRODUCT_ID, 1);
+        $this->addItemToBasket($I, $basketId, self::PRODUCT_ID, 1);
         $this->setBasketDeliveryMethod($I, $basketId, self::SHIPPING_STANDARD);
         $this->setBasketPaymentMethod($I, $basketId, self::PAYMENT_STANDARD);